+++ lib/database/migrations/all_migrations.dart
@@ -35,6 +35,7 @@ import 'package:pantry_app/database/migrations/v42_remove_firebase_cache_meta.da
 import 'package:pantry_app/database/migrations/v43_remove_recipe_shared_id.dart';
 import 'package:pantry_app/database/migrations/v44_query_performance_indexes.dart';
 import 'package:pantry_app/database/migrations/v45_shopping_expiry.dart';
+import 'package:pantry_app/database/migrations/v46_prices_resilience.dart';
 import 'package:pantry_app/database/migrations/v4_nutriscore_grade.dart';
 import 'package:pantry_app/database/migrations/v5_nutriscore_not_applicable.dart';
 import 'package:pantry_app/database/migrations/v6_source_column.dart';
@@ -86,4 +87,5 @@ List<Migration> allMigrations() => [
   MigrationV43(),
   MigrationV44(),
   MigrationV45(),
+  MigrationV46(),
 ];